+++ cmd/neofs-cli/modules/control/object.go
@@ -4,6 +4,8 @@ import (
 	"github.com/spf13/cobra"
 )
 
+const objectFlag = "object"
+
 var objectCmd = &cobra.Command{
 	Use:   "object",
 	Short: "Direct object operations with storage engine",
@@ -12,7 +14,9 @@ var objectCmd = &cobra.Command{
 func initControlObjectsCmd() {
 	objectCmd.AddCommand(listObjectsCmd)
 	objectCmd.AddCommand(objectStatusCmd)
+	objectCmd.AddCommand(reviveObjectCmd)
 
+	initControlObjectReviveCmd()
 	initControlObjectsListCmd()
 	initObjectStatusFlags()
 }